Fix update currency
