--- packages/evershop/src/modules/customer/api/updateProfile/finish[apiResponse].js
@@ -1,12 +1,0 @@
-const { OK } = require('@evershop/evershop/src/lib/util/httpStatus');
-
-// eslint-disable-next-line no-unused-vars
-module.exports = async (request, response, delegate, next) => {
-  const language = await delegate.updateProfile;
-  response.status(OK);
-  response.json({
-    data: {
-      ...language
-    }
-  });
-};
