+++ packages/evershop/src/modules/customer/api/updateProfile/payloadSchema.json
@@ -8,7 +8,7 @@
       "type": "string"
     },
     "email": {
-      "type": "email"
+      "type": "string"
     },
     "language_code": {
       "type": "string"
@@ -17,5 +17,5 @@
       "type": "string"
     }
   },
-  "required": []
+  "additionalProperties": true
 }
